+++ src/components/datagrid/datagrid.js
@@ -4490,6 +4490,22 @@ Datagrid.prototype = {
         }
         return !handle;
       });
+      
+    if (this.toolbar && this.toolbar.parent().find('.table-errors').length > 0) {
+      this.toolbar.parent().find('.table-errors')
+        .off('mouseenter.tableerrortooltip', '.icon')
+        .on('mouseenter.tableerrortooltip', '.icon', function () {
+          handleShow(this);
+        })
+        .off('mouseleave.tableerrortooltip click.tableerrortooltip', '.icon')
+        .on('mouseleave.tableerrortooltip click.tableerrortooltip', '.icon', function () {
+          handleHide(this);
+        })
+        .off('longpress.tableerrortooltip', '.icon')
+        .on('longpress.tableerrortooltip', '.icon', function () {
+          handleShow(this, 0);
+        });
+    }
   },
 
   /**
@@ -8405,12 +8421,19 @@ Datagrid.prototype = {
       this.settings.toolbar = { title: ' ' };
       this.appendToolbar();
     }
-
-    // process via type
-    for (const props in $.fn.validation.ValidationTypes) {  // eslint-disable-line
-      const validationType = $.fn.validation.ValidationTypes[props].type;
-      const errors = $.grep(this.nonVisibleCellErrors, error => error.type === validationType);
-      this.showNonVisibleCellErrorType(errors, validationType);
+    
+    if (this.nonVisibleCellErrors.length === 0) {
+      // remove table-error when not required
+      if (this.toolbar && this.toolbar.parent().find('.table-errors').length === 1) {
+        this.toolbar.parent().find('.table-errors').remove();
+      }    
+    } else {
+      // process via type
+      for (const props in $.fn.validation.ValidationTypes) {  // eslint-disable-line
+        const validationType = $.fn.validation.ValidationTypes[props].type;
+        const errors = $.grep(this.nonVisibleCellErrors, error => error.type === validationType);
+        this.showNonVisibleCellErrorType(errors, validationType);
+      }
     }
   },
 
@@ -8477,6 +8500,7 @@ Datagrid.prototype = {
       isError: type === 'error' || type === 'dirtyerror',
       wrapper: icon
     });
+    this.setupTooltips(false, true);
   },
 
   /**
@@ -8523,6 +8547,22 @@ Datagrid.prototype = {
   clearNonVisibleCellErrors(row, cell, type) {
     if (!this.nonVisibleCellErrors.length) {
       return;
+    }
+
+    if (this.toolbar && this.toolbar.parent() && this.toolbar.parent().find('.table-errors').length > 0) {
+      const icon = this.toolbar.parent().find('.table-errors').find(`.icon-${type}`);
+      if (icon.length) { 
+        const nonVisibleCellTypeErrors = $.grep(this.nonVisibleCellErrors, (error) => {
+          if (error.type === type) {
+            return error;
+          }
+          return '';
+        });
+        // No remaining cell errors of this type
+        if (!nonVisibleCellTypeErrors.length) {
+          icon.remove();
+        }
+      }
     }
 
     this.nonVisibleCellErrors = $.grep(this.nonVisibleCellErrors, (error) => {
@@ -10418,7 +10458,14 @@ Datagrid.prototype = {
       $('body, .scrollable').off('scroll.gridtooltip');
       tooltip.off('touchend.gridtooltip');
       this.element.off('mouseenter.gridtooltip mouseleave.gridtooltip click.gridtooltip longpress.gridtooltip keydown.gridtooltip', selector.str);
-
+      
+      if (this.toolbar && this.toolbar.parent().find('.table-errors').length > 0) {
+        this.toolbar.parent().find('.table-errors')
+          .off('mouseenter.tableerrortooltip', '.icon')
+          .off('mouseleave.tableerrortooltip click.tableerrortooltip', '.icon')
+          .off('longpress.tableerrortooltip', '.icon');
+      }
+      
       // Remove the place component
       const placeApi = tooltip.data('place');
       if (placeApi) {